+++ CHANGELOG.md
@@ -16,3 +16,7 @@
 - Implement `xsd:nonNegativeInteger` data type codec
 - Implement `xsd:long` data type codec
 - Implement `xsd:language` data type codec
+- Implement `xsd:gYearMonth` codec and data type
+- Implement `xsd:gYear` codec and data type
+- Implement `xsd:gMonth` codec and data type
+- Implement `xsd:unsignedShort` data type codec